Added offering and token query parameter parsing [#138854865]
This lets us point an external report from the portal to the dashboard using the /for_class offering url.  The offering is loaded and then the class info url is constructed using the class id found in the first offering for the class (the class id will be the same for all offerings in the response).